Move getMembers from LocalGroup to AndroidGroup
